requirements.txt
langchain
pydantic
langchain-community
langchain-groq
streamlit
python-docx
weasyprint
psycopg2-binary
pdfplumber
